V0.1.1 - Vacation
Most of the last few weeks have been focused on creating models and other placeholder assets.

Implemented a models for crusher and conveyor belt, light changes to how collision checks are performed.
-> Collision visual still isn't working.

Placeholder assets for the fighter are built, just need to work on more machinery to create a functional prototype.